Basic I/O sorted, GUI working, pyudev monitor
Current plan is to have this run as a user service. The pyudev
package is used to monitor for changes to devices; i.e., disc
insert. Then, based on DEVNAME, get the UUID of the disc. If
it exists as JSON file in database, just load data, else,
launch GUI so user can set information for disc. After user
sets all data, the data are saved to file in database.

Next step is to have code that does the actual ripping based on
titles in the info dict. Need to make so that user can supply
only output file generating function as they may want to use the
exact information set in the GUI.