Bug fixes, logging, adds Segments Map
Adds a good deal of logging to whole package.

Info dumps from MakeMKV are stored in compressed gunzip
format.

Adds method for setting window title name to to the name
of the disc that is loaded.

Adds 'Segements Map' info to the stored JSON data for better
matching in future.

Adds 'ripDisc' function for actually ripping disc; is run as
separate process.

Rip title function rips to a 'hidden' directory so that
files being generated with default MakeMKV names are NOT picked
up by the MakeMKV_Watchdog from the video_utils package.

Adds docstrings.