Improves title info storage
Have added some extra keys to the info dictionary for title
information just as source file an segements map. This is done
by trying to .get() these keys from the MakeMKVParser, with
empty strings stored if the keys are not present. E.g., for
DVD, this is no Source FileName key, and for BluRay, there is no
Source Title Id. So, this will be blank depending on which media
is ripped