Adds 'Save & Eject' and 'Save & Rip' buttons
The gui now has two (2) options after defining metadata. The first
is to just save the metadata and eject the disc. No ripping will
occur but the disc database will be updated.

The second option updates the disc database and then ripps the
tracks that have been marked for ripping.