Base changes for multi video/disc
After buying a new 2-film collection, I realized there are
cases where the old setup of disc-wide metadata might not work.
In this case, a single DVD has 2 different movies on it. So, have
reworked the GUI so that the 'disc metadata' widget is ALSO part
of the title/track widget. This way, information can be set
disc-wide OR by track.

Each title in the output JSON file now contains all the old
'disc metadata' as the metadata for different tracks may change.

Need to test the new layout on the aforementioned disc, but must
first update the output file name generate code to use the info
from the titles first.