Disc mounting logic improved, UPC changes
Improved monitoring for mounting discs. When disc mounted would
get a change status for when drive insert/eject, and another
event for mount/unmount. So, when get a state change for disc,
wait for second event of actual mount/unmount until any action
taken.

Have also moved the UPC GUI box to the DiscMetadata class so
that UPC does not appear in each of the title objects