Adds skeleton of system try code
Building app into a systemtray context so that will be
always running and allow user to edit settings without
having to restart program or interact with command line.

Gets 'Quit' working with confirmation dialog. Need to build
a 'settings' dialog that enables pointing to the disc database
and the output file directory for disc backup.