Gets systemtray app working
The systemtry application is now fully functional, including
a settings menu to set database location, output location, and
what types of titles to rip.

Need to add a check for the output directory on launch. If the
output directory is NOT found, then need to launch dialog/warning
for user to set new output directory.

Alos want to add a user interact timeout dialog when disc is
is inserted. The idea is that when a disc is inserted that already
exists in the database, a dialog will appear for 30(?) seconds
asking if the user wants to update the data for the disc OR just
rip the disc. If there is no interaction within the timeout period,
will assume user just wants to rip the disc. Otherwise, launch the
metadata editing application.