Changes multiprocess start method
Attempts to fix some collison bugs by moving to a spawn start
context instead of fork. Hopefully this will keep things a little
more separated. Had to add some code for logging from different
processes and a thread to read logs from logging queue and place
in main logger.

Also makes modification to the 'mounting' list so that drives that
are mounting a disc are not added multiple times. That is, if disc
is NOT mounted, but is in in the mounting list, it is not added to
mounting list again. Multiple adds to the mounting list may have
also been causing errors as not all instances of the 'dev' string
are removed with the .remove() method.