Threads for ripping and status
After some playing, realized that there are some bugs/shortfalls
in getting a progress bar to display in the TrayIcon, at least on
Ubunut.

So, moved to a check box in the system tray to enable/disable
ripping status. The ripping status will be displayed in a separate
dialog, with a 'widget' added for each disc drive that is ripping.

Still trying to figure out how to get the ripper thread(s) to
send signals to the dialog to show progress. Also, not sure how
to initalize the dialog in the main thread using signals such that
it is only open when somthing is ripping; i.e., shows up when
a rip starts and closes when a rip finishes.