Reworks QThreads for MakeMKV
The way the QThreads were layed out for running the makemkvcon
command was not great. So, reworked to have a "base" MakeMKVThread
class (a subclass of QThread) that is responsible for building/
executing the makemkvcon command using the makemkvcon() method.

There is now a MakeMKVInfo class and a MakeMKVRip class, with the
former responsible for getting title information for the mediainfo
GUI and the latter responsible for actually ripping the titles.

Both of these class are subclasses of the MakeMKVThread calls,
with overloaded run() methods that call the makemkvcon() method
to start their respective MakeMKV jobs and then parse the output
accordingly.

All this was done so that can have a more unified way of kill the
threads through the QThread.quit() method. The MakeMKVThread class
overloads the quit() method so that the makemkvcon subprocess
(if any) is killed before calling super().quit() to kill the thread.

This can be cascaded up to any other threads (e.g., the ripper
watchdog) so that there is a consistent way to kill off threads.