Fixes multi-disc rip bugs
Lots of bugs when trying to rip multiple discs. These steamed from
guessing at which discs were to be ripped after metadata read/entered
by user. Mitigated this by overloading the done() method of QDialog
and adding a new FINISHED signal so that the dev device could be passed
around to track what device was being worked on. Also did this for
the progress bar so that could cancel correct rip. This also involved
adding a check in the Ripper class' kill() method where check if the
device passed in matches the one being worked on.

Updates how name of package is passed to objects; i.e., there is
a new NAME variable in the top-level __init__ file that is used.
This is done in preperation of 'combining' with the cdripper code.