Adds titles table to existing dialog
When the "existing disc" dialog came up, there was no way to tell
what information in the database it matched to. So, borrowed some
code from the cdripper to add a table of titles that were found
in the library for the disc so that the user can decide if the
info matches.