Overhauls progress and gets backup working
Large overhaul of the ripping progress bars to use information output
by makemkv directly instead of scaning file size. This required changing
signal names and updating some methods for the progress widget.
Signal names were modified to be different than those of the cdripper
package so that could all be combined into the autoripper package.

Updates to makemkv.py so that the base MakeMKV class is a bit more
flexible in how it can be called/run.

When only one title to be ripped, then run the 'old' way of ripping
single file from disc. If more than one, then we make a decrypted
backup of the disc, then extract each of the titles from the backup.
This gets rid of overhead of having to scan the disc each time a
title is to be ripped.