Attempts to add complete status, updates progress
Adds some signals and MakeMKV message parsing in attempt to have
pop-up to inform user of success/failure during ripping. The message
parsing seems to work as get log message with status, but the signal
is not triggered up the stack fully. Need to investigate further.
It is possible that the object sending the signal is garbage
collected before fully sent? Not sure.

Also updates progress widget a litte for disc backups so that the
extraction progress of individual tracks from the backup image is
displayed in the progress widget.