Fixes bug with mixed discs
Some discs have both a movie and series episodes on them
(e.g., Thomas the Tank Engine DVDs). For these discs, only those
titles that match the disc type (i.e., Movie or Series) would be
extracted.

This has been fixed by iterating of the titles of the disc instead
of taking the overall disc type as the type for all titles.