Bug fixes for Blu-Ray backup/extract
Did not realize that when backing up a Blu-Ray disc, a directory of
files is created, not an ISO file like DVD backup. What's more, when
scaning the backup to extract titles, titles are not necessarily added
in the same order as when the optical media is scanned.

To get around this, the source file name from the JSON file is used,
along with mkvmerge from MKVToolNix, to extract the titles from the
backup. This lead to the creation of two (2) new methods for a class
in the ripper.py module; one for extracting titles from DVD backup
that is the same as the old method of extraction and a new method
for exctraction titles from the directory backup using mkvmerge.
Which method runs depends on the the `media_type` value in the JSON
information.

Updates to rip progress have also been made so that the mkvmerge
progress can be displayed. Metadata for each title extraction are
now also displayed as the titles are pulled from the backup.