Bug fixes for metadata and progress widgets
After updating the progress widgets, introduced a bug where was not
passing a 'pipe' argument to the NEW_PROCESS signal when new disc
inserted. Also had a bug with the typehint for the new_process()
method on the BasicWidget class; had comma instead of colon for hint.

Updates the docstring for the new_process() method to explain arguments.