Fixes makemkv parser bug
There was a major bug introduced when overhauling the progress
widgets of the code for full disc rips, namely blu-ray.

The issue was that status messages from makemkvcon are sent to
stdout, while progress messages are sent to stderr. However,
I failed to properly specify that stderr was to be used in the
progress widget, causing it to start grabbing and parsing
lines from stdout.

This caused titles to failed to display, or be lost during the
scan of the disc, causing very odd behavior.

I have added locking to the ProgressParser class and moved the
pipe and proc attributes to be read-only properties that are
updated through the new update_proc_pipe() method. These variables
are now all forced thread safe through a threading.Lock object.