Modifies disc eject sequence
In previous versions, the disc was not ejected during a full disc
rip until after all the files were extracted from the back up.
This was a big of a waste because it can take quite a bit of time
depending on how many files to extract from the backup. So, the
call to the 'eject' command has been moved into the MakeMKVRip
class with a check to see if the source for the command is a
dev or disc object before trying to run 'eject'. This way, after
a disc is finished being backed up, it will eject while files are
being extracted from the backup, allowing for a new disc to be
inserted and saving some time.