Implements new disc hashing
Moves to using new disc hashing method in hopes that can get
code working on windows/macOS. This includes migration of database
when on linux system to new hashing.

Will have to create new watch dog for use on windows machines,
along with other modifications for use of dev in disc handling...
will probably have to move that to watchdog so disc handler
can stay a little more general.