Chages disc_hash.py to match TheDiscDB
TheDiscDB is an online database containing information about which
titles are a disc to extract. Thanks to the amount of data that
is stored in the DB, and the open-source code, I was able to update
the disc hashing code here to match. This means that as discs are
added the the autoMakeMKV database, their hashes will match that
of the TheDiscDB.

I have also built a new thediscdb_utils.py module that can be used
to convert data from TheDiscDB into valid autoMakeMKV files. This
feature has not been fully tested, but have some more work, and
getting this new hashing fully implemented into the code, I plan
to cross link this project with TheDiscDB.

Note I also have to test autoMakeMKV on Windows and get it working
there as the ImportBuddy (TheDiscDB ingest program) code is in C#.