Builds out plex file namer
Finally built out a file namer for Plex format. Need to add into
the settings UI to select between the video_utils and plex naming
styles. Need to also add boxes to set the name of the movie and
series libraries for the plex namer.

The output naming modules have been moved to a new paths/
subpackage with video_utils.py and plex.py modules. Each of these
modules contains a movie() and series() function for handling
movies and series, respectively. IN the paths/__init__.py file,
then is an outfile() function defined to get output file names.
The style= keyword can be set to 'video_utils' or 'plex' to switch
between the two.