Fix plex convention bug, updates API
Have fixed a bug in the plex movie output file naming where the
movie directory was not being created. Still have to test this more
for both movies and series for an 'all titles' extract.

Have updated the outfile() api to chagne the 'style' keyword to
'convention' for naming convention.