Updates to settings dialog
The settings dialog is now comprised of a Dialog and a Widget. This
is done so that the widget, which holds the actual code and layout,
can be reused in autoripper in a tab under that settings dialog.

Have also add a QComboBox for setting the naming convention; have to
test that this works properly