Improves hashing and migrate
Noticed that on a disc, there was an issue getting file sizes that
caused crash. Have moved file size checking to separate function
with try/except and retries for getting file sizes to (hopefully)
eliminate this issue.

Have updated DB migration a little for injection of thediscid into
the json files when moving them