Gets working on windows
To get running on windows, adds new global varibles/attrs to the
main __init__.py module for MAKEMKVCON and MKVMERGE. The shutil.which()
function is used to search for the two CLIs, with search params
for linux and windows machines. Have also updated APPDIR to be
machine specific.

Bug fix in the disc watchdogs where the HANDLE_DISC signal was
attached to a method twice; in the base class and in subclass. Now
is only attached in the base class.

Updates directory checking at beginning to check for both outdir
and dbdir. While this works, there is a bug further down the
pipe where the dbdir value is not being honored. It does show up
in the setting.json file though.