Fixes tmp dir bug
On linux, the disc dev is /dev/srN, where N is a number. Was using
the srN part to name a temporary directory for output files.
However, windows just gives the drive letter (e.g., E:), which
gives no 'basename', and so no temporary directory was being created.
Now using the os.path.splitdrive() function to get drive name and
use that (if it exists) as the temporary directory name.