Updates message/signal handling
Have made some improvements to code logic for makemkv. We now check
both the return code of the subprocess and the 'failure' property
of the class to see if we should continue on in the code.

The rip() method also now returns a boolean value indicating
success (True) or Failure (False). On a success, then we try
to remove the temporary directory; on failure, is removed in
code for proper cleanup of all files.

Small changes to logging order to better indicate what is happening