Removes eject from MakeMKV, signaling
The call to eject the disc has been removed from the makemkv code
in favor of doing it elsewhere with signals/slots.

Have add a CANCEL signal, but having issues using it as realized
the creation of MakeMKVThread objects within the Ripper QThread
is likely messing with signal connections. Need to rework how
all this happens...