Modifies signals and eject handling
Added signals and changed method names to help improve ejection
of disc handling. It is possible that a disc could be ejected
while the ripper/disc handler object is still working, so we now
add the object to the _mounted and _running dicts, with the object
removed from _running on a call to FINISHED from the dischandler
object. The object is added to the list using the disc hash so
that other discs won't overwrite it.