Moves disc_hash to QThread object
As the disc hashing can take some time and lock-up the GUI, have
converted this to a QThread object. Now, when a DiscHandler object
is created, the disc hashing is started and its FINISHED signal
is connected to the disc_lookup() method so that the GUI remains
responsive while hashing occurs and the disc lookup is triggered
after hashing is complete.