Reorganization of class/flow
Have moved much of the attribute definitions from the __init__
method of the windows and linux watchdogs into the base watchdog
to simplify things.

The windows dialog now just uses the progress widget as the handler
for the disc events instead of creating a seperate widget.

Have updated the linux watchdog a bit to (hopefully) account for
all the programatic changes with signaling; will need to test