Improves load_metadata, log updates
In previous versions of code, would check the output directory
during rip to get progress and so would need to get the size of
each title from the MakeMKV output. However, have since updated
to directly use the MakeMKV status information for progress
tracking. So, I have removed the title size extraction code from
the function, which has sped it up a bit.

Updates logging in the db_migrate() function