Adds backup then tag event loop
Ran into some Blu-ray discs that were hard to determine which
title was which due to lengths begin very/exactly similar.
Thus, a new 'Backup Then Tag' button has been added to the
metadata editor that runs a backup of the disc, then reopens
the metadata editor for tagging.

This allows the user to determine which track, based on STREAM
names, is which and then tag the disc. When selecting the 'Rip'
button after this, titles are simply ripped from the existing
backup; in old method, would have had to back up the disc twice.

Also fixes bug in SUCCESS and FAILURE signals in the ExtractFromBluRay.

Fixes bug in table generation for the options dialog when disc
is found in database