Delayed eject for backup then tag
In the "Backup Then Tag" pipeline, we now delay eject of disc until
after the tagging is done; i.e., disc backup, re-open metadata
dialog, select rip, then eject. This is done so that the disc can
be loaded into a play (e.g., VLC) to determine which file is which.

Also adds a check of the mkvmerge output for success/failure flag.

Increases dev to mount point timeout from 20 to 60 seconds on linux.