Fixes bug with hashid and tests code
Realized that the hashid attribute of the DiscHandler class was
used to store the disc hash or the DiscHasher object. This was,
for obvious reasons, causing some issues. So, I created a new
disc_hasher attribute to store the DiscHasher() object, with hashid
holding the hash values.

Removes the raise Exception calls placed for testing of program flow