TheDiscDb Zip archive support, metadata bug fix
Can now load/read from a zip archive of TheDiscDb downloaded from
Github. This way don't have to unpack the whole thing, can just do
in RAM/on-the-fly while converting the autoMakeMKV format.

Adds fix to the Metadata Editor when launched from systemtray so
that MakeMKV does not run for disc scan as the widget is not
associated with a disc when launching from system tray.