Mount point bug fix; Samba testing
There was a bug for a disc where the path in /dev/disc/by-label
had escaped unicode space characters that, when building path
to mount point, would not be valid on linux (ubuntu) system.
Have added a secondary loop to try the mount path as-is, with a
second try escaping the string.

Adds a cleanup attribute to the DiscHandler class to test mkvextract
for files on Samba share. It seems there is just an issue with
mkvextract running on files on a network drive, at least on linux.

Ups the number of attempts and sleep time on the DiscHasher class
to improve chances of getting hash for disc.