Existing backup/Backup then tag (#6)
* Gets success/failure dialogs working

The backup/extract/rip success and failure dialogs are now working
with dialogs that inform of success automatically closing after 30
seconds and error dialogs stay "forever".

To do this, the MyQDialog class was reworked to handle the auto
closing. This also means a slight rework of the ExistingDiscOptions
class

* Adds backup then tag event loop

Ran into some Blu-ray discs that were hard to determine which
title was which due to lengths begin very/exactly similar.
Thus, a new 'Backup Then Tag' button has been added to the
metadata editor that runs a backup of the disc, then reopens
the metadata editor for tagging.

This allows the user to determine which track, based on STREAM
names, is which and then tag the disc. When selecting the 'Rip'
button after this, titles are simply ripped from the existing
backup; in old method, would have had to back up the disc twice.

Also fixes bug in SUCCESS and FAILURE signals in the ExtractFromBluRay.

Fixes bug in table generation for the options dialog when disc
is found in database

* Finshed/Fail for extract, mount lookup thread

Have added SIGNALS to the ExtractFromBluray class for SUCCESS
and FAILURE so now can trigger a success or failure dialog for
each titl extraction.

Have moved the mount point lookup code to a QThread so that happens
in background. The FINISHED signal is connected to new
found_mount_point() slot, passing in the mount point. This new
method/slot sets up the DiscHash thread to compute hash same as in
previous version.

* Delayed eject for backup then tag

In the "Backup Then Tag" pipeline, we now delay eject of disc until
after the tagging is done; i.e., disc backup, re-open metadata
dialog, select rip, then eject. This is done so that the disc can
be loaded into a play (e.g., VLC) to determine which file is which.

Also adds a check of the mkvmerge output for success/failure flag.

Increases dev to mount point timeout from 20 to 60 seconds on linux.

* Working to support extract from existing

It is possible that a user may have an existing backup of a disc
and want to just extract from that. So, working to update to have
more consistent backup/tmpdir names along with new dialog to
see how user wants to proceed if existing backup is found.

* Fixes bug with hashid and tests code

Realized that the hashid attribute of the DiscHandler class was
used to store the disc hash or the DiscHasher object. This was,
for obvious reasons, causing some issues. So, I created a new
disc_hasher attribute to store the DiscHasher() object, with hashid
holding the hash values.

Removes the raise Exception calls placed for testing of program flow

* Adds Metadata Editor option to SysTray

Have added a 'button' to launch the Metadata Editor widget to
the system tray. This way can updated/edit metadata information

* Updates TheDiscDb to check for editions

The editions of movies is stored in the title of the track in
TheDiscDb archive. Have added a check and a user prompt to see
if possible edition information is, in fact, edition information.

* TheDiscDb Zip archive support, metadata bug fix

Can now load/read from a zip archive of TheDiscDb downloaded from
Github. This way don't have to unpack the whole thing, can just do
in RAM/on-the-fly while converting the autoMakeMKV format.

Adds fix to the Metadata Editor when launched from systemtray so
that MakeMKV does not run for disc scan as the widget is not
associated with a disc when launching from system tray.

* Mount point bug fix; Samba testing

There was a bug for a disc where the path in /dev/disc/by-label
had escaped unicode space characters that, when building path
to mount point, would not be valid on linux (ubuntu) system.
Have added a secondary loop to try the mount path as-is, with a
second try escaping the string.

Adds a cleanup attribute to the DiscHandler class to test mkvextract
for files on Samba share. It seems there is just an issue with
mkvextract running on files on a network drive, at least on linux.

Ups the number of attempts and sleep time on the DiscHasher class
to improve chances of getting hash for disc.