Adds title select to existing disc dialog
When a disc that already exists in the data base is inserted,
a dialog was presented asking the user what to do (e.g., rip,
ignore, etc.). There was no way to tell the program which titles
to rip, with title selection handled by options in the settings
menu.

Now, the options in the settings menu set the default titles to rip,
but the user can override/changed what to rip in the existing disc
dialog. This means that when the timed dialog opens, only those titles
that meet the default criteria (Features only, extras only, everything)
will be automatically selected in the dialog.

Program flow has also been changed a bit so that the existing disc
dialog is also presented after tagging metadata IF a rip of the
disc is requested. This allows the user to select specific tracks
after tagging the dics for the first time.