Adds support for temporary directory
When running a backup and extract, it may be nice to back up the
disc to one location, but extract titles to another. So, have added
a temporary directory entry field in the settings dialog. By default,
this matches the output directory, but can be changed by user. Disc
backups will go the temporary directory, with extracted titles
going to the output directory.

Have updated the metdata widget a little to maintain title selctions
when doing a backup then tag pipeline. This was done by moving the
code that did the title information parsing in the save() method to
its own method that is now called in both the save() method and
the backup_then_tag() method to save title information (if any)
into the JSON file.

Have added a counter for failed extractions from a disc backup by
checking that the previous_output variables is not empty and that
the path not exists for a failed counter. If the counter is > 0,
then we do NOT remove the backup incase we want to try again.