Updates signaling for bluray extract, refactor
Have updated the ProgessDialog to have MKV_PROGRESS_TITLE and
MKV_PROGRESS_VALUE signals that can be used to update the progress
bar titles and values, respectively, from the main progress dialog.
This makes the code in the ExtractFromBluray a bit cleaner as can
just call these signals and it handles finding the corresponding
widget and everything.

Restructured the progress.py module a bit so the class were grouped
better (less important at bottom), with the ProgressWidget being
moved to the top as it is used in the return typehint for the new
get_widget method of the ProgressDialog.