Adds code for remote FS check
Code to check if path/filesystem is remote file. This matters
because mkvmerge does not seem to like remote systems. So, for
backup then extract on bluray, need to backup to local, extract
to local, then move to remote through python.

Need to implement check of outdir path on boot and settings change
to see if is remote path. If is, then we need to set a tmpdir using
the tempfile python package. Otherwise, tmpdir will just be outdir.