Moves user settings to Settings class
The user settings (set in the SettingsDialog) are now managed by
a Settings class defined in settings.py. This is done so that we
don't need to pass around settings anymore.

The main Settings object is initailzed in the __init__.py file
for the packages, with all other modules importing this object.
The settings dialog updates the attributes of the 'global' object,
with all other code getting their settings informaiton from the object.

As I write this, I realize that the settings in the object are currently
updated as the user changes them, with a cancel not actually canceling
any changes; need to change that...

Remove user specified temporary directory; is determined for them.

Updates full backup location to only be forced to local if output dir
is remote and is blu-ray disc, otherwise just use output dir.