Settings changes are not cancelled
Have added a cancel method to the Settings class that will quietly
reload the settings file; no remote directory check warning is thrown.

In the SettingsDialog, have added a changed attribute that is set
to True if any of the connected methods (for change events) is
called. This way, if settings is opened, then closed, nothing
happens; no need to reload settings.