Removes saving of info from makemkv (#11)
* Adds support for temporary directory

When running a backup and extract, it may be nice to back up the
disc to one location, but extract titles to another. So, have added
a temporary directory entry field in the settings dialog. By default,
this matches the output directory, but can be changed by user. Disc
backups will go the temporary directory, with extracted titles
going to the output directory.

Have updated the metdata widget a little to maintain title selctions
when doing a backup then tag pipeline. This was done by moving the
code that did the title information parsing in the save() method to
its own method that is now called in both the save() method and
the backup_then_tag() method to save title information (if any)
into the JSON file.

Have added a counter for failed extractions from a disc backup by
checking that the previous_output variables is not empty and that
the path not exists for a failed counter. If the counter is > 0,
then we do NOT remove the backup incase we want to try again.

* Get temp dir 'fully' working

For bluray backup, mkvmerge just does not like samba shares. So
the temporary directory has been implemented. The tmp dir should
be a local system so that things are extracted there and then
placed into the final destination. For single title extraction
or DVD, direct to output directory should work fine.

Also gets progress bar labels working of mkvmerge and for moving
of data file via python; moved in chunks with progress tracked.

Makes DevToMount a litte more robust by wrapping calls in try/except
for edge case I found were root directy was not found when trying
to list the directory.

Fixes bug in title extract from bluray where clicking Cancel on
progress widget would crash the program due to missing signal

* Updates signaling for bluray extract, refactor

Have updated the ProgessDialog to have MKV_PROGRESS_TITLE and
MKV_PROGRESS_VALUE signals that can be used to update the progress
bar titles and values, respectively, from the main progress dialog.
This makes the code in the ExtractFromBluray a bit cleaner as can
just call these signals and it handles finding the corresponding
widget and everything.

Restructured the progress.py module a bit so the class were grouped
better (less important at bottom), with the ProgressWidget being
moved to the top as it is used in the return typehint for the new
get_widget method of the ProgressDialog.

* Adds code for remote FS check

Code to check if path/filesystem is remote file. This matters
because mkvmerge does not seem to like remote systems. So, for
backup then extract on bluray, need to backup to local, extract
to local, then move to remote through python.

Need to implement check of outdir path on boot and settings change
to see if is remote path. If is, then we need to set a tmpdir using
the tempfile python package. Otherwise, tmpdir will just be outdir.

* Adds --no-cleanup CLI flag

When set, any disc backups will not be deleted after extraction
of titles is complete. By default, will remove files.

* Moves user settings to Settings class

The user settings (set in the SettingsDialog) are now managed by
a Settings class defined in settings.py. This is done so that we
don't need to pass around settings anymore.

The main Settings object is initailzed in the __init__.py file
for the packages, with all other modules importing this object.
The settings dialog updates the attributes of the 'global' object,
with all other code getting their settings informaiton from the object.

As I write this, I realize that the settings in the object are currently
updated as the user changes them, with a cancel not actually canceling
any changes; need to change that...

Remove user specified temporary directory; is determined for them.

Updates full backup location to only be forced to local if output dir
is remote and is blu-ray disc, otherwise just use output dir.

* Settings changes are not cancelled

Have added a cancel method to the Settings class that will quietly
reload the settings file; no remote directory check warning is thrown.

In the SettingsDialog, have added a changed attribute that is set
to True if any of the connected methods (for change events) is
called. This way, if settings is opened, then closed, nothing
happens; no need to reload settings.

* Removes MakeMKV info store, metadata edit

To keep the database size smaller, have decided to remove storing of
the MakeMKV output (.info.gz) files. This changes some things about
work flow. Will now no longer be able to edit data without inserting
a disc to get the disc handler dialog with 'Open' option to appear.
You can still manually edit JSON files, but would not recommend.
Also, when doing this, will have to rescan disc to get MakeMKV info.

To do:
For backup then edit, need to maintain the information from the first scan
so that don't have to rescan (again) when tagging.

* Backup then tag maintain data fix

Now maintains the MakeMKV information from initial scan when running
back then tag. The data are stored in the DiscHandler class after
the initial scan and passed into new keyword arguments for the
metadata editor class. When passed in, these bypass running scan
of disc and are just used directly for metadata tree build.