Bug fix for QThread close failure
Was encountering an issue when running code in autoripper program
where QThreads were being closed improperly and crashing the program.

To address this, emitting of SUCCESS and FAILURE signals have been
moved into the run() methods of the ripper classes so that they
are emitted as late as possible, or as close to closing them as
possible. In the previous version, these signals could be emitted
before trying to clean up any files, which means the thread could
have been doing things while trying to clean them up.

I have also added sender.join() calls to the failed, success, and
finished methods in the base watchdog class before trying to do
anything else. This is done to ensure the that the thread is done
before cleaning anything up.