Changes join() to wait()
There is no join() method on QThread objects, instead is wait. Also,
as the signal from the ripper/extractor classes are passed through
the DiscHandler class, the sender object to the finish, success, failure
methods is a DiscHandler object. So, added a wait() method to disc
handler that checks/waits for ripper or extractor threads to finish