Fix for QThread progress kill (#15)
Bug in the Metadata editor where cleaned up the BasicProgressWidget
before its QThread was fully done. Overloaded the wait() method
of the ProgressParser thread to just 'forget' the subprocess it
was parsing and then call super().wait().

Have also added a wait() method to the BasicProgessWidget what waits
for the parser thread. This method is now called in the metadata
editor before cleaning up